Cleanup of recipe logic (#6883)
