+++ src/main/java/gregtech/api/recipe/lookup/GTItemDataLookupIngredient.java
@@ -1,8 +1,12 @@
 package gregtech.api.recipe.lookup;
 
 import java.util.Objects;
+import java.util.function.Consumer;
+
+import net.minecraft.item.ItemStack;
 
 import gregtech.api.objects.ItemData;
+import gregtech.api.util.GTOreDictUnificator;
 
 public final class GTItemDataLookupIngredient extends GTRecipeLookupIngredient {
 
@@ -23,6 +27,14 @@ public final class GTItemDataLookupIngredient extends GTRecipeLookupIngredient {
             throw new IllegalArgumentException("itemData unification name must not be empty");
         }
         return new GTItemDataLookupIngredient(unificationName);
+    }
+
+    public static void fromRuntime(Consumer<? super GTItemDataLookupIngredient> ingredients, ItemStack stack) {
+        ItemData itemData = GTOreDictUnificator.getAssociation(stack);
+        if (itemData == null) return;
+        String unificationName = itemData.toString();
+        if (unificationName.isEmpty()) return;
+        ingredients.accept(new GTItemDataLookupIngredient(unificationName));
     }
 
     @Override
