+++ src/main/java/gregtech/api/recipe/maps/NACRecipeMapBackend.java
@@ -2,12 +2,17 @@ package gregtech.api.recipe.maps;
 
 import java.util.Collection;
 
+import javax.annotation.ParametersAreNonnullByDefault;
+
 import org.jetbrains.annotations.NotNull;
 
 import gregtech.api.recipe.RecipeMapBackend;
 import gregtech.api.recipe.RecipeMapBackendPropertiesBuilder;
 import gregtech.api.util.GTRecipe;
+import gregtech.api.util.MethodsReturnNonnullByDefault;
 
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
 public class NACRecipeMapBackend extends RecipeMapBackend {
 
     private int maxDuration;
