+++ src/test/java/gregtech/api/recipe/RecipeMapBackendLookupTest.java
@@ -18,9 +18,12 @@ import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Spliterators;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
+
+import javax.annotation.ParametersAreNonnullByDefault;
 
 import net.minecraft.init.Items;
 import net.minecraft.item.Item;
@@ -39,15 +42,19 @@ import cpw.mods.fml.common.registry.RegistryDelegate;
 import gregtech.api.enums.Materials;
 import gregtech.api.objects.ItemData;
 import gregtech.api.recipe.lookup.GTFluidLookupIngredient;
+import gregtech.api.recipe.lookup.GTItemDataLookupIngredient;
 import gregtech.api.recipe.lookup.GTItemStackLookupIngredient;
 import gregtech.api.recipe.lookup.GTRecipeLookup;
 import gregtech.api.recipe.lookup.GTRecipeLookupIngredient;
 import gregtech.api.util.GTLog;
 import gregtech.api.util.GTOreDictUnificator;
 import gregtech.api.util.GTRecipe;
+import gregtech.api.util.MethodsReturnNonnullByDefault;
 import sun.misc.Unsafe;
 import sun.reflect.ReflectionFactory;
 
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
 class RecipeMapBackendLookupTest {
 
     private static final Constructor<GTRecipe> GT_RECIPE_CONSTRUCTOR = constructorFor(GTRecipe.class);
@@ -276,7 +283,7 @@ class RecipeMapBackendLookupTest {
 
         IllegalStateException error = assertThrows(
             IllegalStateException.class,
-            () -> RecipeMapBackend.validateLookup(mapName, backend));
+            () -> RecipeLookupValidator.validateLookup(mapName, backend));
 
         assertTrue(
             error.getMessage()
@@ -306,7 +313,7 @@ class RecipeMapBackendLookupTest {
 
         IllegalStateException error = assertThrows(
             IllegalStateException.class,
-            () -> RecipeMapBackend.validateLookup(mapName, backend));
+            () -> RecipeLookupValidator.validateLookup(mapName, backend));
 
         assertTrue(
             error.getMessage()
@@ -333,7 +340,7 @@ class RecipeMapBackendLookupTest {
         backend.compileRecipe(secondRegistered);
         backend.compileRecipe(thirdRegistered);
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup(mapName, backend));
+        assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup(mapName, backend));
     }
 
     @Test
@@ -356,7 +363,7 @@ class RecipeMapBackendLookupTest {
         backend.compileRecipe(exactRecipe);
         backend.compileRecipe(wildcardRecipe);
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup(mapName, backend));
+        assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup(mapName, backend));
     }
 
     @Test
@@ -372,7 +379,8 @@ class RecipeMapBackendLookupTest {
         FilterRejectingLookupBackend backend = new FilterRejectingLookupBackend(recipe);
         backend.compileRecipe(recipe);
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.wildcard_rejected", backend));
+        assertDoesNotThrow(
+            () -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.wildcard_rejected", backend));
     }
 
     @Test
@@ -396,7 +404,7 @@ class RecipeMapBackendLookupTest {
         backend.compileRecipe(firstRecipe);
         backend.compileRecipe(queryRecipe);
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.nbt_seed", backend));
+        assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.nbt_seed", backend));
     }
 
     @Test
@@ -413,7 +421,7 @@ class RecipeMapBackendLookupTest {
 
         IllegalStateException error = assertThrows(
             IllegalStateException.class,
-            () -> RecipeMapBackend.validateLookup(mapName, backend));
+            () -> RecipeLookupValidator.validateLookup(mapName, backend));
 
         assertTrue(
             error.getMessage()
@@ -436,7 +444,7 @@ class RecipeMapBackendLookupTest {
 
         IllegalStateException error = assertThrows(
             IllegalStateException.class,
-            () -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.short_circuit", backend));
+            () -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.short_circuit", backend));
 
         assertTrue(
             error.getMessage()
@@ -458,7 +466,7 @@ class RecipeMapBackendLookupTest {
         FilterRejectingLookupBackend backend = new FilterRejectingLookupBackend(recipe);
         backend.compileRecipe(recipe);
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.filter_reject", backend));
+        assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.filter_reject", backend));
     }
 
     @Test
@@ -468,7 +476,7 @@ class RecipeMapBackendLookupTest {
         backend.compileRecipe(
             recipe(item("lookup.validation.unindexed.input"), item("lookup.validation.unindexed.output"), category));
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.unindexed", backend));
+        assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.unindexed", backend));
     }
 
     @Test
@@ -507,7 +515,7 @@ class RecipeMapBackendLookupTest {
                     category));
 
             assertDoesNotThrow(
-                () -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.cross_unification", backend));
+                () -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.cross_unification", backend));
         } finally {
             if (hadPreviousTarget) {
                 GTOreDictUnificator.getName2StackMap()
@@ -558,7 +566,7 @@ class RecipeMapBackendLookupTest {
                     item("lookup.validation.item_data.equivalent.output"),
                     category));
 
-            assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.item_data", backend));
+            assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.item_data", backend));
         } finally {
             if (hadPreviousTarget) {
                 GTOreDictUnificator.getName2StackMap()
@@ -579,7 +587,7 @@ class RecipeMapBackendLookupTest {
         RecipeCategory category = allocate(RECIPE_CATEGORY_CONSTRUCTOR);
         backend.compileRecipe(recipeWithoutInputs(item("lookup.validation.empty.output"), category));
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.empty", backend));
+        assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.empty", backend));
     }
 
     @Test
@@ -590,7 +598,7 @@ class RecipeMapBackendLookupTest {
         recipe.mFakeRecipe = true;
         backend.compileRecipe(recipe);
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.fake", backend));
+        assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.fake", backend));
     }
 
     @Test
@@ -604,7 +612,7 @@ class RecipeMapBackendLookupTest {
         recipe.mEnabled = false;
         backend.compileRecipe(recipe);
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.disabled", backend));
+        assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.disabled", backend));
     }
 
     @Test
@@ -621,7 +629,7 @@ class RecipeMapBackendLookupTest {
 
         IllegalStateException error = assertThrows(
             IllegalStateException.class,
-            () -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.min_input_gate", backend));
+            () -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.min_input_gate", backend));
         assertTrue(
             error.getMessage()
                 .contains("query rejected before trie lookup"));
@@ -649,7 +657,7 @@ class RecipeMapBackendLookupTest {
 
         IllegalStateException error = assertThrows(
             IllegalStateException.class,
-            () -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.missing_oredict", backend));
+            () -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.missing_oredict", backend));
         assertTrue(
             error.getMessage()
                 .contains("unresolved oredict input(s)=1"));
@@ -684,7 +692,7 @@ class RecipeMapBackendLookupTest {
             GTOreDictUnificator.resetUnificationEntries();
 
             List<GTRecipeLookupIngredient> group = new ArrayList<>();
-            RecipeMapBackend.addRuntimeItemStackLookupIngredients(group, equivalent);
+            GTItemStackLookupIngredient.fromRuntime(group::add, equivalent);
 
             assertTrue(group.contains(GTItemStackLookupIngredient.fromRuntime(equivalent)));
             assertTrue(group.contains(GTItemStackLookupIngredient.fromRuntimeWildcard(equivalent)));
@@ -725,7 +733,7 @@ class RecipeMapBackendLookupTest {
         backend.compileRecipe(nullableSlotRecipe);
         backend.compileRecipe(concreteRecipe);
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.nullable", backend));
+        assertDoesNotThrow(() -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.nullable", backend));
     }
 
     @Test
@@ -747,7 +755,8 @@ class RecipeMapBackendLookupTest {
 
         backend.compileRecipe(recipe);
 
-        assertDoesNotThrow(() -> RecipeMapBackend.validateLookup("gt.recipe.lookup.test.alt_representative", backend));
+        assertDoesNotThrow(
+            () -> RecipeLookupValidator.validateLookup("gt.recipe.lookup.test.alt_representative", backend));
     }
 
     private static GTRecipe recipe(Item input, Item output, RecipeCategory category) {
@@ -915,17 +924,19 @@ class RecipeMapBackendLookupTest {
         protected Stream<GTRecipe> lookupCandidateStream(@Nullable ItemStack @NotNull [] items,
             @Nullable FluidStack @NotNull [] fluids) {
             List<GTRecipeLookupIngredient> ingredients = new ArrayList<>();
+            Consumer<GTRecipeLookupIngredient> adder = ingredients::add;
 
             for (ItemStack item : items) {
                 if (item == null) continue;
 
-                RecipeMapBackend.addRuntimeItemStackLookupIngredients(ingredients, item);
+                GTItemStackLookupIngredient.fromRuntime(adder, item);
+                GTItemDataLookupIngredient.fromRuntime(adder, item);
             }
 
             for (FluidStack fluid : fluids) {
                 if (fluid == null || fluid.getFluid() == null) continue;
 
-                ingredients.add(new GTFluidLookupIngredient(fluid));
+                GTFluidLookupIngredient.fromRuntime(adder, fluid);
             }
 
             if (ingredients.isEmpty()) {
